requirements.txt

reflex==0.8.13
gspread==5.7.2
python_dotenv==1.1.1
google_genai==1.48.0
pillow==12.0.0
